Feature/sim watermark metrics

+++ build-all.go
@@ -7,9 +7,11 @@ package main
 
 import (
 	"./github.com/ostelco-core/goscript"
+	"flag"
 	"fmt"
 	"log"
 	"os"
+	"os/signal"
 )
 
 func generateEspEndpointCertificates(originalCertPath string, activeCertPath string, domainName string) {
@@ -81,7 +83,19 @@ func generateDummyStripeEndpointSecretIfNotSet() {
 }
 
 func main() {
+
+	cleanPtr := flag.Bool("clean", false, "If set, run a './gradlew clean' before building and testing.")
+	stayUpPtr := flag.Bool("stay-up", false, "If set, keep test environment up after running tests.")
+	flag.Parse()
+
 	log.Printf("About to get started\n")
+	if *cleanPtr {
+		log.Printf("    ... will clean.")
+	}
+
+	if *stayUpPtr {
+		log.Printf("    ... will keep environment up after acceptance tests have run.")
+	}
 
 	//
 	// Ensure that  all preconditions for building and testing are met, if not
@@ -104,14 +118,41 @@ func main() {
 		"ocsgw/cert/metrics.crt",
 		"ocs.dev.ostelco.org")
 
+	buildUsingGradlew(cleanPtr)
+
+	runIntegrationTestsViaDocker(stayUpPtr)
+
+	log.Printf("Build and integration tests succeeded\n")
+}
+
+func runIntegrationTestsViaDocker(stayUpPtr *bool) {
+
+	// First take down any lingering docker jobs that may interfer with
+	// the current run.
+	goscript.AssertSuccesfulRun("docker-compose down")
+
+	if *stayUpPtr {
+		// If  ctrl-c is pushed during stay-up, then take the whole thing down using docker-compose down
+		c := make(chan os.Signal, 1)
+		signal.Notify(c, os.Interrupt)
+		go func() {
+			<-c
+			goscript.AssertSuccesfulRun("docker-compose down")
+			os.Exit(1)
+		}()
+		goscript.AssertSuccesfulRun("docker-compose up --build")
+	} else {
+		goscript.AssertSuccesfulRun("docker-compose up --build --abort-on-container-exit")
+	}
+}
+
+func buildUsingGradlew(cleanPtr *bool) {
 	//
 	// All preconditions are now satisfied, now run the actual build/test commands
 	// and terminate the build process if any of them fails.
 	//
-
+	if *cleanPtr {
+		goscript.AssertSuccesfulRun("./gradlew build")
+	}
 	goscript.AssertSuccesfulRun("./gradlew build")
-	goscript.AssertSuccesfulRun("docker-compose down")
-	goscript.AssertSuccesfulRun("docker-compose up --build --abort-on-container-exit")
-
-	log.Printf("Build and integration tests succeeded\n")
 }
